[cleanup] Add logic to iterate on compression and delete files as needed.
Previously we'd attempt to compress a minimal amount of files to stay under the
maxCacheSize limit, but we wouldn't verify that compression actually got us
there. Now we iterate repeatedly as needed to ensure that we stay under the
maxCacheSize limit.  Additionally if we cannot compress enough files to stay
under the passed limit, we'll start deleting files.